src/main/java/com/anirudhology/leetcode/tree/DiameterOfBinaryTree.java
package com.anirudhology.leetcode.tree;

import com.anirudhology.leetcode.util.TreeNode;

public class DiameterOfBinaryTree {

    private int diameter = 0;

    public int diameterOfBinaryTree(TreeNode root) {
        // Find maximum depth and update diameter
        maxDepth(root);
        return diameter;
    }

    private int maxDepth(TreeNode root) {
        if (root == null) {
            return 0;
        }
        // Calculate heights of the left and right subtrees
        int leftHeight = maxDepth(root.left);
        int rightHeight = maxDepth(root.right);
        // Update diameter
        diameter = Math.max(diameter, leftHeight + rightHeight);
        return 1 + Math.max(leftHeight, rightHeight);
    }
}
